Moved SqlFileStream class, then moved the PInvoke files which it depended upon
