+++ src/Microsoft.Data.SqlClient/src/Common/CoreLib/Interop/Windows/kernel32/Interop.SetThreadErrorMode.cs
