+++ src/Microsoft.Data.SqlClient/src/Common/Interop/Windows/Interop.Errors.cs
