+++ src/Microsoft.Data.SqlClient/src/Common/Interop/Windows/Interop.Libraries.cs
