+++ src/Microsoft.Data.SqlClient/src/Common/Interop/Windows/Interop.UNICODE_STRING.cs
