+++ src/Microsoft.Data.SqlClient/src/Common/Interop/Windows/NtDll/Interop.FILE_FULL_EA_INFORMATION.cs
