+++ src/Microsoft.Data.SqlClient/src/Common/Interop/Windows/NtDll/Interop.IO_STATUS_BLOCK.cs
