+++ src/Microsoft.Data.SqlClient/src/Common/Interop/Windows/NtDll/Interop.RtlNtStatusToDosError.cs
