+++ src/Microsoft.Data.SqlClient/src/Common/Interop/Windows/kernel32/Interop.CTL_CODE.cs
