+++ src/Microsoft.Data.SqlClient/src/Common/Interop/Windows/kernel32/Interop.DeviceIoControl.cs
