+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlTypes/SqlFileStream.Unsupported.cs
